fix: skip hot.accept() when a non-component export exists
For example, if a module exports both a React component and a React hook,
changes to the React hook must bubble up to any dependent components.
For this to happen, the module containing the changed hook cannot call
hot.accept, or else the dependent components won't rerender as they should.